Move load user logic into redirectUserToDefaultTeam
