+++ components/mfa/confirm.jsx
@@ -3,7 +3,6 @@
 
 import React from 'react';
 import {FormattedMessage} from 'react-intl';
-import PropTypes from 'prop-types';
 
 import Constants from 'utils/constants.jsx';
 import {isKeyPressed} from 'utils/utils.jsx';
@@ -15,13 +14,6 @@ import FormattedMarkdownMessage from 'components/formatted_markdown_message.jsx'
 const KeyCodes = Constants.KeyCodes;
 
 export default class Confirm extends React.Component {
-    static propTypes = {
-        shouldLoadUser: PropTypes.bool,
-        actions: PropTypes.shape({
-            loadMe: PropTypes.func.isRequired,
-        }),
-    }
-
     componentDidMount() {
         document.body.addEventListener('keydown', this.onKeyPress);
     }
@@ -32,12 +24,7 @@ export default class Confirm extends React.Component {
 
     submit = (e) => {
         e.preventDefault();
-
-        if (this.props.shouldLoadUser) {
-            this.props.actions.loadMe().then(() => redirectUserToDefaultTeam());
-        } else {
-            redirectUserToDefaultTeam();
-        }
+        redirectUserToDefaultTeam();
     }
 
     onKeyPress = (e) => {
